requirements.txt
pandas==1.3.5
python-docx==0.8.11
